Update to upstream version '4.0.12'
with Debian dir 7faca2f34003d4a6eb54630255ad192b185a93b3

+++ jstests/aggregation/bugs/server3253.js
@@ -1,3 +1,10 @@
+// Cannot implicitly shard accessed collections because unsupported use of sharded collection
+// for output collection of aggregation pipeline.
+// @tags: [
+//   assumes_superuser_permissions,
+//   assumes_unsharded_collection,
+// ]
+
 // server-3253 Unsharded support for $out
 load('jstests/aggregation/extras/utils.js');
 
@@ -43,7 +50,7 @@ function test(input, pipeline, expected) {
 
 function listCollections(name) {
     var collectionInfosCursor = db.runCommand("listCollections", {filter: {name: name}});
-    return new DBCommandCursor(db.getMongo(), collectionInfosCursor).toArray();
+    return new DBCommandCursor(db, collectionInfosCursor).toArray();
 }
 
 input.insert({_id: 1});
@@ -87,10 +94,12 @@ test(input, [{$project: {c: {$concat: ["hello there ", "_id"]}}}], [
     {_id: 3, c: "hello there _id"}
 ]);
 
-// test with capped collection
-cappedOutput.drop();
-db.createCollection(cappedOutput.getName(), {capped: true, size: 2});
-assertErrorCode(input, {$out: cappedOutput.getName()}, 17152);
+// test with capped collection, skip for mobile SE as it doesn't support capped collections.
+if (jsTest.options().storageEngine !== "mobile") {
+    cappedOutput.drop();
+    db.createCollection(cappedOutput.getName(), {capped: true, size: 2});
+    assertErrorCode(input, {$out: cappedOutput.getName()}, 17152);
+}
 
 // ensure everything works even if input doesn't exist.
 test(inputDoesntExist, [], []);
