+++ src/mongo/shell/shell_utils_launcher.cpp
@@ -1,30 +1,32 @@
 // mongo/shell/shell_utils_launcher.cpp
-/*
- *    Copyright 2010 10gen Inc.
+
+/**
+ *    Copyright (C) 2018-present MongoDB, Inc.
  *
- *    This program is free software: you can redistribute it and/or  modify
- *    it under the terms of the GNU Affero General Public License, version 3,
- *    as published by the Free Software Foundation.
+ *    This program is free software: you can redistribute it and/or modify
+ *    it under the terms of the Server Side Public License, version 1,
+ *    as published by MongoDB, Inc.
  *
  *    This program is distributed in the hope that it will be useful,
  *    but WITHOUT ANY WARRANTY; without even the implied warranty of
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *    GNU Affero General Public License for more details.
+ *    Server Side Public License for more details.
  *
- *    You should have received a copy of the GNU Affero General Public License
- *    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *    You should have received a copy of the Server Side Public License
+ *    along with this program. If not, see
+ *    <http://www.mongodb.com/licensing/server-side-public-license>.
  *
  *    As a special exception, the copyright holders give permission to link the
  *    code of portions of this program with the OpenSSL library under certain
  *    conditions as described in each individual source file and distribute
  *    linked combinations including the program with the OpenSSL library. You
- *    must comply with the GNU Affero General Public License in all respects
- *    for all of the code used other than as permitted herein. If you modify
- *    file(s) with this exception, you may extend this exception to your
- *    version of the file(s), but you are not obligated to do so. If you do not
- *    wish to do so, delete this exception statement from your version. If you
- *    delete this exception statement from all source files in the program,
- *    then also delete it in the license file.
+ *    must comply with the Server Side Public License in all respects for
+ *    all of the code used other than as permitted herein. If you modify file(s)
+ *    with this exception, you may extend this exception to your version of the
+ *    file(s), but you are not obligated to do so. If you do not wish to do so,
+ *    delete this exception statement from your version. If you delete this
+ *    exception statement from all source files in the program, then also delete
+ *    it in the license file.
  */
 
 #define MONGO_LOG_DEFAULT_COMPONENT ::mongo::logger::LogComponent::kDefault
@@ -78,7 +80,6 @@ extern char** environ;
 
 namespace mongo {
 
-using std::unique_ptr;
 using std::cout;
 using std::endl;
 using std::make_pair;
@@ -86,6 +87,7 @@ using std::map;
 using std::pair;
 using std::string;
 using std::stringstream;
+using std::unique_ptr;
 using std::vector;
 
 #ifdef _WIN32
@@ -259,25 +261,33 @@ void ProgramOutputMultiplexer::clear() {
     _buffer.str("");
 }
 
-ProgramRunner::ProgramRunner(const BSONObj& args, const BSONObj& env) {
-    verify(!args.isEmpty());
+ProgramRunner::ProgramRunner(const BSONObj& args, const BSONObj& env, bool isMongo) {
+    uassert(ErrorCodes::FailedToParse,
+            "cannot pass an empty argument to ProgramRunner",
+            !args.isEmpty());
 
     string program(args.firstElement().valuestrsafe());
-    verify(!program.empty());
+    uassert(ErrorCodes::FailedToParse,
+            "invalid program name passed to ProgramRunner",
+            !program.empty());
     boost::filesystem::path programPath = findProgram(program);
+    boost::filesystem::path programName = programPath.stem();
+
 
     string prefix("mongod-");
-    bool isMongodProgram =
-        string("mongod") == program || program.compare(0, prefix.size(), prefix) == 0;
+    bool isMongodProgram = isMongo && (string("mongod") == programName ||
+                                       programName.string().compare(0, prefix.size(), prefix) == 0);
     prefix = "mongos-";
-    bool isMongosProgram =
-        string("mongos") == program || program.compare(0, prefix.size(), prefix) == 0;
+    bool isMongosProgram = isMongo && (string("mongos") == programName ||
+                                       programName.string().compare(0, prefix.size(), prefix) == 0);
 
-    if (isMongodProgram) {
+    if (!isMongo) {
+        _name = "sh";
+    } else if (isMongodProgram) {
         _name = "d";
     } else if (isMongosProgram) {
         _name = "s";
-    } else if (program == "mongobridge") {
+    } else if (programName == "mongobridge") {
         _name = "b";
     } else {
         _name = "sh";
@@ -290,6 +300,7 @@ ProgramRunner::ProgramRunner(const BSONObj& args, const BSONObj& env) {
     // Parse individual arguments into _argv
     BSONObjIterator j(args);
     j.next();  // skip program name (handled above)
+
     while (j.more()) {
         BSONElement e = j.next();
         string str;
@@ -301,12 +312,14 @@ ProgramRunner::ProgramRunner(const BSONObj& args, const BSONObj& env) {
             verify(e.type() == mongo::String);
             str = e.valuestr();
         }
-        if (str == "--port") {
-            _port = -2;
-        } else if (_port == -2) {
-            _port = strtol(str.c_str(), 0, 10);
-        } else if (isMongodProgram && str == "--configsvr") {
-            _name = "c";
+        if (isMongo) {
+            if (str == "--port") {
+                _port = -2;
+            } else if (_port == -2) {
+                _port = strtol(str.c_str(), 0, 10);
+            } else if (isMongodProgram && str == "--configsvr") {
+                _name = "c";
+            }
         }
         _argv.push_back(str);
     }
@@ -356,10 +369,12 @@ ProgramRunner::ProgramRunner(const BSONObj& args, const BSONObj& env) {
         ++environEntry;
     }
 #endif
-    bool needsPort = isMongodProgram || isMongosProgram || (program == "mongobridge");
+    bool needsPort =
+        isMongo && (isMongodProgram || isMongosProgram || (programName == "mongobridge"));
     if (!needsPort) {
         _port = -1;
     }
+
     uassert(ErrorCodes::FailedToParse,
             str::stream() << "a port number is expected when running " << program
                           << " from the shell",
@@ -474,6 +489,11 @@ boost::filesystem::path ProgramRunner::findProgram(const string& prog) {
         p = prog + ".exe";
     }
 #endif
+
+    // The file could exist if it is specified as a full path.
+    if (p.is_absolute() && boost::filesystem::exists(p)) {
+        return p;
+    }
 
     // Check if the binary exists in the current working directory
     boost::filesystem::path t = boost::filesystem::current_path() / p;
@@ -735,8 +755,12 @@ BSONObj ClearRawMongoProgramOutput(const BSONObj& args, void* data) {
 
 BSONObj CheckProgram(const BSONObj& args, void* data) {
     ProcessId pid = ProcessId::fromNative(singleArg(args).numberInt());
-    bool isDead = wait_for_pid(pid, false);
-    return BSON(string("") << (!isDead));
+    int exit_code = -123456;  // sentinel value
+    bool isDead = wait_for_pid(pid, false, &exit_code);
+    if (!isDead) {
+        return BSON("" << BSON("alive" << true));
+    }
+    return BSON("" << BSON("alive" << false << "exitCode" << exit_code));
 }
 
 BSONObj WaitProgram(const BSONObj& a, void* data) {
@@ -772,7 +796,7 @@ BSONObj StartMongoProgram(const BSONObj& a, void* data) {
         }
     }
 
-    ProgramRunner r(args, env);
+    ProgramRunner r(args, env, true);
     r.start();
     invariant(registry.isPidRegistered(r.pid()));
     stdx::thread t(r);
@@ -780,9 +804,9 @@ BSONObj StartMongoProgram(const BSONObj& a, void* data) {
     return BSON(string("") << r.pid().asLongLong());
 }
 
-BSONObj RunMongoProgram(const BSONObj& a, void* data) {
+BSONObj RunProgram(const BSONObj& a, void* data, bool isMongo) {
     BSONObj env{};
-    ProgramRunner r(a, env);
+    ProgramRunner r(a, env, isMongo);
     r.start();
     invariant(registry.isPidRegistered(r.pid()));
     stdx::thread t(r);
@@ -792,10 +816,21 @@ BSONObj RunMongoProgram(const BSONObj& a, void* data) {
     return BSON(string("") << exit_code);
 }
 
+BSONObj RunMongoProgram(const BSONObj& a, void* data) {
+    return RunProgram(a, data, true);
+}
+
+BSONObj RunNonMongoProgram(const BSONObj& a, void* data) {
+    return RunProgram(a, data, false);
+}
+
 BSONObj ResetDbpath(const BSONObj& a, void* data) {
     verify(a.nFields() == 1);
     string path = a.firstElement().valuestrsafe();
-    verify(!path.empty());
+    if (path.empty()) {
+        warning() << "ResetDbpath(): nothing to do, path was empty";
+        return undefinedReturn;
+    }
     if (boost::filesystem::exists(path))
         boost::filesystem::remove_all(path);
     boost::filesystem::create_directory(path);
@@ -805,7 +840,10 @@ BSONObj ResetDbpath(const BSONObj& a, void* data) {
 BSONObj PathExists(const BSONObj& a, void* data) {
     verify(a.nFields() == 1);
     string path = a.firstElement().valuestrsafe();
-    verify(!path.empty());
+    if (path.empty()) {
+        warning() << "PathExists(): path was empty";
+        return BSON(string("") << false);
+    };
     bool exists = boost::filesystem::exists(path);
     return BSON(string("") << exists);
 }
@@ -842,8 +880,10 @@ BSONObj CopyDbpath(const BSONObj& a, void* data) {
     BSONObjIterator i(a);
     string from = i.next().str();
     string to = i.next().str();
-    verify(!from.empty());
-    verify(!to.empty());
+    if (from.empty() || to.empty()) {
+        warning() << "CopyDbpath(): nothing to do, source or destination path(s) were empty";
+        return undefinedReturn;
+    }
     if (boost::filesystem::exists(to))
         boost::filesystem::remove_all(to);
     boost::filesystem::create_directory(to);
@@ -992,8 +1032,8 @@ BSONObj getStopMongodOpts(const BSONObj& a) {
 /** stopMongoProgram(port[, signal]) */
 BSONObj StopMongoProgram(const BSONObj& a, void* data) {
     int nFields = a.nFields();
-    verify(nFields >= 1 && nFields <= 3);
-    uassert(15853, "stopMongo needs a number", a.firstElement().isNumber());
+    uassert(ErrorCodes::FailedToParse, "wrong number of arguments", nFields >= 1 && nFields <= 3);
+    uassert(ErrorCodes::BadValue, "stopMongoProgram needs a number", a.firstElement().isNumber());
     int port = int(a.firstElement().number());
     int code = killDb(port, ProcessId::fromNative(0), getSignal(a), getStopMongodOpts(a));
     log() << "shell: stopped mongo program on port " << port;
@@ -1001,11 +1041,13 @@ BSONObj StopMongoProgram(const BSONObj& a, void* data) {
 }
 
 BSONObj StopMongoProgramByPid(const BSONObj& a, void* data) {
-    verify(a.nFields() == 1 || a.nFields() == 2);
-    uassert(15852, "stopMongoByPid needs a number", a.firstElement().isNumber());
+    int nFields = a.nFields();
+    uassert(ErrorCodes::FailedToParse, "wrong number of arguments", nFields >= 1 && nFields <= 3);
+    uassert(
+        ErrorCodes::BadValue, "stopMongoProgramByPid needs a number", a.firstElement().isNumber());
     ProcessId pid = ProcessId::fromNative(int(a.firstElement().number()));
-    int code = killDb(0, pid, getSignal(a));
-    log() << "shell: stopped mongo program on pid " << pid;
+    int code = killDb(0, pid, getSignal(a), getStopMongodOpts(a));
+    log() << "shell: stopped mongo program with pid " << pid;
     return BSON("" << (double)code);
 }
 
@@ -1017,10 +1059,28 @@ int KillMongoProgramInstances() {
         int port = registry.portForPid(pid);
         int code = killDb(port != -1 ? port : 0, pid, SIGTERM);
         if (code != EXIT_SUCCESS) {
+            log() << "Process with pid " << pid << " exited with error code " << code;
             returnCode = code;
         }
     }
     return returnCode;
+}
+
+std::vector<ProcessId> getRunningMongoChildProcessIds() {
+    std::vector<ProcessId> registeredPids, outPids;
+    registry.getRegisteredPids(registeredPids);
+    // Only return processes that are still alive. A client may have started a program using a mongo
+    // helper but terminated another way. E.g. if a mongod is started with MongoRunner.startMongod
+    // but exited with db.shutdownServer.
+    std::copy_if(registeredPids.begin(),
+                 registeredPids.end(),
+                 std::back_inserter(outPids),
+                 [](const ProcessId& pid) {
+                     const bool block = false;
+                     bool isDead = wait_for_pid(pid, block);
+                     return !isDead;
+                 });
+    return outPids;
 }
 
 MongoProgramScope::~MongoProgramScope() {
@@ -1032,6 +1092,7 @@ void installShellUtilsLauncher(Scope& scope) {
     scope.injectNative("runProgram", RunMongoProgram);
     scope.injectNative("run", RunMongoProgram);
     scope.injectNative("_runMongoProgram", RunMongoProgram);
+    scope.injectNative("runNonMongoProgram", RunNonMongoProgram);
     scope.injectNative("_stopMongoProgram", StopMongoProgram);
     scope.injectNative("stopMongoProgramByPid", StopMongoProgramByPid);
     scope.injectNative("rawMongoProgramOutput", RawMongoProgramOutput);
@@ -1042,5 +1103,5 @@ void installShellUtilsLauncher(Scope& scope) {
     scope.injectNative("pathExists", PathExists);
     scope.injectNative("copyDbpath", CopyDbpath);
 }
-}
-}
+}  // namespace shell_utils
+}  // namespace mongo
